fix(class)[develop]: backend - add use case for check restrictions of specific class.
